[qspi_dbi] Rename from qspi_amoled, add features (#7594)
