+++ esphome/const.py
@@ -398,6 +398,7 @@ CONF_INCLUDES = "includes"
 CONF_INDEX = "index"
 CONF_INDOOR = "indoor"
 CONF_INFRARED = "infrared"
+CONF_INIT_SEQUENCE = "init_sequence"
 CONF_INITIAL_MODE = "initial_mode"
 CONF_INITIAL_OPTION = "initial_option"
 CONF_INITIAL_STATE = "initial_state"
